Rename stuff as preparation to consensus blocks
